project/requirements.txt
aiofiles==23.1.0
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
fastapi==0.95.0
filelock==3.13.1
flower==1.2.0
fsspec==2023.12.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.20.3
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.2
kombu==5.3.5
MarkupSafe==2.1.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
openai==1.10.0
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
pydantic==1.10.14
pydantic_core==2.16.1
pytest==7.2.2
python-multipart==0.0.7
pytz==2024.1
PyYAML==6.0.1
redis==4.5.4
regex==2023.12.25
requests==2.28.2
rfc3986==1.5.0
safetensors==0.4.2
sniffio==1.3.0
starlette==0.26.1
sympy==1.12
tokenizers==0.15.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
transformers==4.37.2
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.21.1
validators==0.22.0
vine==5.1.0
wcwidth==0.2.13
